version bump and imgui changes

+++ OOBlugin.cs
@@ -12,7 +12,7 @@ using System.Linq;
 using Dalamud.Game;
 using Dalamud.Plugin;
 using Dalamud.Plugin.Services;
-using ImGuiNET;
+using Dalamud.Bindings.ImGui;
 
 namespace OOBlugin
 {
@@ -215,7 +215,7 @@ namespace OOBlugin
         [HelpMessage("Toggles New Game+.")]
         private unsafe void OnNGPT(string command, string argument)
         {
-            Game.newGameUIPtr = (Game.newGameUIPtr != nint.Zero) ? Game.newGameUIPtr : DalamudApi.GameGui.FindAgentInterface("QuestRedo") + 0xD0;
+            Game.newGameUIPtr = (nint)((Game.newGameUIPtr != nint.Zero) ? Game.newGameUIPtr : DalamudApi.GameGui.FindAgentInterface("QuestRedo") + 0xD0);
             if (Game.newGameUIPtr == nint.Zero) { PrintError("Failed to get NG+ agent, please open the NG+ window and then use this command to initialize it."); return; }
 
             *(byte*)(Game.newGameUIPtr + 0x8) ^= 1;
